tools/ansible-vault-keys/pyproject.toml
[project]
name = "ansible-vault-keys"
version = "0.1.0"
description = "Selective encryption of YAML keys using Ansible Vault, with full round-trip support"
authors = [
    { name = "Joe DeKeyrel", email = "joedekeyrel@email.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "ansible>=2.18",
    "ruamel.yaml>=0.18",
]

[project.scripts]
ansible-vault-keys = "encrypt_values:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
